Moved content builder to different namespace
