+++ src/TransferGenerator/Definition/DefinitionFactory.php
@@ -9,15 +9,14 @@ use Picamator\TransferObject\Shared\SharedFactoryTrait;
 use Picamator\TransferObject\TransferGenerator\Config\ConfigFactoryTrait;
 use Picamator\TransferObject\TransferGenerator\Definition\Filesystem\DefinitionFinder;
 use Picamator\TransferObject\TransferGenerator\Definition\Filesystem\DefinitionFinderInterface;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Builder\ContentBuilder;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Builder\ContentBuilderInterface;
 use Picamator\TransferObject\TransferGenerator\Definition\Parser\Builder\EmbeddedTypeBuilder;
 use Picamator\TransferObject\TransferGenerator\Definition\Parser\Builder\EmbeddedTypeBuilderInterface;
 use Picamator\TransferObject\TransferGenerator\Definition\Parser\Builder\NamespaceBuilder;
 use Picamator\TransferObject\TransferGenerator\Definition\Parser\Builder\NamespaceBuilderInterface;
+use Picamator\TransferObject\TransferGenerator\Definition\Parser\ContentBuilder;
+use Picamator\TransferObject\TransferGenerator\Definition\Parser\ContentBuilderInterface;
 use Picamator\TransferObject\TransferGenerator\Definition\Parser\DefinitionParser;
 use Picamator\TransferObject\TransferGenerator\Definition\Parser\DefinitionParserInterface;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\TypePropertyExpander;
 use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\CollectionTypePropertyExpander;
 use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\DateTimeTypePropertyExpander;
 use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\EnumTypePropertyExpander;
@@ -25,6 +24,7 @@ use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\Nullab
 use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\NumberTypePropertyExpander;
 use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\PropertyExpanderInterface;
 use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\ProtectedPropertyExpander;
+use Picamator\TransferObject\TransferGenerator\Definition\Parser\Expander\TypePropertyExpander;
 use Picamator\TransferObject\TransferGenerator\Definition\Reader\DefinitionReader;
 use Picamator\TransferObject\TransferGenerator\Definition\Reader\DefinitionReaderInterface;
 use Picamator\TransferObject\TransferGenerator\Definition\Validator\Content\ClassNameContentValidator;
