+++ src/TransferGenerator/Definition/Parser/DefinitionParser.php
@@ -7,7 +7,6 @@ namespace Picamator\TransferObject\TransferGenerator\Definition\Parser;
 use Generator;
 use Picamator\TransferObject\Dependency\YmlParser\YmlParserInterface;
 use Picamator\TransferObject\TransferGenerator\Definition\Enum\BuildInTypeEnum;
-use Picamator\TransferObject\TransferGenerator\Definition\Parser\Builder\ContentBuilderInterface;
 
 readonly class DefinitionParser implements DefinitionParserInterface
 {
